+++ ui/src/components/NavBar.tsx
@@ -1,7 +1,11 @@
 import { useState } from 'react'
-import { NavBarProps } from '../types/navbar.ts'
 import PathBar from './PathBar'
 import useWindowStore from '../state/useWindowStore.ts'
+
+export interface NavBarProps {
+  id: number
+  path: string
+}
 
 export default function NavBar({ id, path }: NavBarProps) {
   const [hovered, setHovered] = useState(false)
