--- ui/src/types/api.ts
@@ -1,6 +1,0 @@
-export interface HTTPRequest {
-  url: string
-  method: string
-  headers: Record<string, string>
-  body: string
-}
