--- ui/src/types/navbar.ts
@@ -1,4 +1,0 @@
-export interface NavBarProps {
-  id: number
-  path: string
-}
