--- ui/src/types/sigil.ts
@@ -1,8 +1,0 @@
-export interface SigilProps {
-  point: string
-  size: string
-  detail: string
-  space: string
-  background: string
-  foreground: string
-}
